LPD-17786 announcements-web: Simplify
