+++ modules/apps/announcements/announcements-web/src/main/java/com/liferay/announcements/web/internal/display/context/AnnouncementsAdminViewManagementToolbarDisplayContext.java
@@ -59,8 +59,7 @@ public class AnnouncementsAdminViewManagementToolbarDisplayContext {
 		_renderRequest = renderRequest;
 
 		_announcementsAdminViewDisplayContext =
-			new DefaultAnnouncementsAdminViewDisplayContext(
-				_httpServletRequest);
+			new AnnouncementsAdminViewDisplayContext(_httpServletRequest);
 		_currentURLObj = PortletURLUtil.getCurrent(
 			liferayPortletRequest, liferayPortletResponse);
 		_themeDisplay = (ThemeDisplay)httpServletRequest.getAttribute(
